+++ components/NoticePage/NoticeItem/style.ts
